Upgrade data without table recreating

+++ src/com/kozzztya/cycletraining/db/datasources/ExercisesDS.java
@@ -4,6 +4,7 @@ import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
+
 import com.kozzztya.cycletraining.db.DBHelper;
 import com.kozzztya.cycletraining.db.entities.Exercise;
 
@@ -35,10 +36,11 @@ public class ExercisesDS extends DataSource<Exercise> {
 
     public static void onUpgrade(SQLiteDatabase database, int oldVersion,
                                  int newVersion) {
-        Log.v(DBHelper.LOG_TAG, "Upgrading table " + TABLE_NAME + " from version "
-                + oldVersion + " to " + newVersion);
-        database.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-        onCreate(database);
+        //Recreate table if it was created before stable version
+        if (oldVersion <= DBHelper.DATABASE_VERSION_STABLE) {
+            database.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(database);
+        }
     }
 
     @Override
